some more fixes

+++ model/common/src/icon4py/model/common/metrics/compute_diffusion_metrics.py
@@ -71,7 +71,7 @@ def _compute_ls_params(
     z_maxhgtd_avg: xp.ndarray,
     c_owner_mask: xp.ndarray,
     thslp_zdiffu: float,
-    thhgtd_zdiffu: int,
+    thhgtd_zdiffu: float,
     cell_nudging: int,
     n_cells: int,
     nlev: int,
